feat: Display Notes app name in the bar
